+++ src/main/java/np/edu/nast/payroll/Payroll/repository/EmployeeSalaryComponentRepository.java
@@ -1,4 +1,4 @@
 package np.edu.nast.payroll.Payroll.repository;
 import np.edu.nast.payroll.Payroll.entity.EmployeeSalaryComponent;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface EmployeeSalaryComponentRepository extends JpaRepository<EmployeeSalaryComponent, Long> {}
+public interface EmployeeSalaryComponentRepository extends JpaRepository<EmployeeSalaryComponent, Integer> {}
