+++ src/main/java/np/edu/nast/payroll/Payroll/service/EmployeeLeaveService.java
@@ -5,10 +5,16 @@ import np.edu.nast.payroll.Payroll.entity.EmployeeLeave;
 import java.util.List;
 
 public interface EmployeeLeaveService {
+
     EmployeeLeave requestLeave(EmployeeLeave leave);
-    EmployeeLeave updateLeave(Long id, EmployeeLeave leave);
-    void deleteLeave(Long id);
-    EmployeeLeave getLeaveById(Long id);
+
+    EmployeeLeave updateLeave(Integer id, EmployeeLeave leave);
+
+    void deleteLeave(Integer id);
+
+    EmployeeLeave getLeaveById(Integer id);
+
     List<EmployeeLeave> getAllLeaves();
-    List<EmployeeLeave> getLeavesByEmployee(Long empId);
+
+    List<EmployeeLeave> getLeavesByEmployee(Integer empId);
 }
